+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 4.1.0
+ July 17, 2023
+- Up-to-date with plugin, SharedPreferences:
+  static bool setPrefix(String prefix, {Set<String>? allowList}) {
+- sdk: '>=2.18.0 <4.0.0'
+
 ## 4.0.0
  December 02, 2022
 - static dynamic get(String? key) to static Object? get(String? key)